+++ tests/test_unit/conftest.py
@@ -1,24 +1,23 @@
 import json
 from collections.abc import Callable
-from typing import TYPE_CHECKING, TypeAlias
+from typing import TypeAlias
 
 import pytest
 from django.conf import LazySettings
 from django.http import HttpResponse
 from inline_snapshot import snapshot
 
-if TYPE_CHECKING:
-    CsrfFailureAssertion: TypeAlias = Callable[[HttpResponse], None]
+CsrfFailureAssertion: TypeAlias = Callable[[HttpResponse], None]
 
 
 @pytest.fixture(
-    params=[(True), (False)],
+    params=[True, False],
     ids=['debug_on', 'debug_off'],
 )
 def assert_csrf_failure_message(
     settings: LazySettings,
     request: pytest.FixtureRequest,
-) -> 'CsrfFailureAssertion':
+) -> CsrfFailureAssertion:
     """Assert CSRF failure message according to debug mode setting."""
     is_debug_mode_active = request.param
     settings.DEBUG = is_debug_mode_active